Throw GattClosed if Gatt is closed during I/O operation (#23)
Typical behavior of `Channel` is to throw a
`ClosedReceiveChannelException` on invocation of `close` if a `receive`
is suspended. This is unfortunately a bit counter intuitive and can
produce misleading stacktraces (i.e. stacktrace will include `close`
invocation but not `receive`).

In order to be more explicit with our exception stacktraces related to
`Channel` cancelation: catching `ClosedReceiveChannelException` and
explicitly throwing `GattClosed` exception (that carries original
`ClosedReceiveChannelException` as it's cause).

May need to revisit this behavior in the future after `receiveOrClosed`
is implemented per Kotlin/kotlinx.coroutines#330.