Remove FailedToDeliverEvent Exception
Was thrown if we failed to `offer` response in `GattCallback`, but
`onResponse` is a `CONFLATED` `Channel` so will never return `false`.

We're protected by the `Mutex` in `CoroutinesGatt` which enforces that
requests are executed sequentially, so we should never have a response
trample a previous (response is received from `onResponse` `Channel`
prior to the next request going out).