Add characteristic change BroadcastChannel
`Flow`'s `onStart` is called **after** subscribing to underlying data
stream. When performing a send-then-receive the response could be lost
if the `Flow` is slow to subscribe.

Problem was described/discussed in [Kotlin/kotlinx.coroutines#1758] and
will be fixed in [Kotlin/kotlinx.coroutines#2034].

The workaround is to use `openSubscription().consumeAsFlow()` which
opens the subscription prior to spinning up the `Flow` (and in turn will
be subscribed before `onStart`).

[Kotlin/kotlinx.coroutines#1758]: https://github.com/Kotlin/kotlinx.coroutines/issues/1758#issuecomment-626372261
[Kotlin/kotlinx.coroutines#2034]: https://github.com/Kotlin/kotlinx.coroutines/issues/2034